+++ src/types/index.ts
@@ -4,8 +4,8 @@ export { DistributionMethod } from '../sub-clients/common/model/DistributionMeth
 export { Link } from '../sub-clients/common/model/LinkEntity';
 export { AdminLinks } from '../sub-clients/adminLinks/model/AdminLinks';
 export { Collection } from '../sub-clients/collections/model/Collection';
-export { ContentPartner } from '../sub-clients/contentPartners/model/ContentPartner';
-export { ContentPartnerRequest } from '../sub-clients/contentPartners/model/ContentPartnerRequest';
+export { Channel } from '../sub-clients/channels/model/Channel';
+export { ContentPartnerRequest } from '../sub-clients/channels/model/ContentPartnerRequest';
 export { LegalRestriction } from '../sub-clients/legalRestrictions/model/LegalRestriction';
 export { Subject } from '../sub-clients/subjects/model/Subject';
 export { Order } from '../sub-clients/orders/model/Order';