+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.0.0] - 2020-06-01
+
+- Rename ContentPartnersClient -> ChannelsClient
+- Rename ContentPartner -> Client
+
 ## [10.7.0] - 2020-05-30
 
 - Allow platformInteractedWith events to be sent anonymously
@@ -22,15 +27,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `sort_by` to VideoSearchRequest
 
 ## [10.5.2] - 2020-05-18
+
 - Remove `adminCollectionSearch` use `searchCollections` instead
 - Add `getMyCollections` for convenience
 
 ## [10.5.1] - 2020-05-18
+
 - Rename `public` to `discoverable` (in collections)
 
 ## [10.5.0] - 2020-05-18
 
-- Add url to collectionInteractedWith event to set boclips-referer as normal referer is unreliable 
+- Add url to collectionInteractedWith event to set boclips-referer as normal referer is unreliable
 - Add `contentWarnings` subclient, update admin links to get all content warnings
 - Update video request can contain content warnings